+++ src/aks-preview/azext_aks_preview/_params.py
@@ -1416,6 +1416,9 @@ def load_arguments(self, _):
             help="delete an AKS nodepool by ignoring PodDisruptionBudget setting",
         )
 
+    with self.argument_context("aks nodepool delete-machines") as c:
+        c.argument("--machine-names", help="machines to delete.")
+
     with self.argument_context("aks machine") as c:
         c.argument("cluster_name", help="The cluster name.")
         c.argument(